Endpoint de clientes optimizados
